# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts API",
    "description": "Documentation generated by swagger-autogen."
  },
  "host": "https://jatsi341-assignment3.onrender.com/",
  "basePath": "/",
  "tags": [
    {
      "name": "Contacts",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/contacts/": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Endpoint to get all contacts.",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "description": "Endpoint to create a new contact.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The contact was created successfully"
          },
          "500": {
            "description": "Faling creating the contact"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "description": "Endpoint to get a single contact.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact ID."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "description": "Endpoint to update a contact.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact ID."
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The contact was updated successfully"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Faling updating the contact"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "description": "Endpoint to delete a contact.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Contact ID."
          }
        ],
        "responses": {
          "201": {
            "description": "The contact was deleted successfully"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Faling deleting the contact"
          }
        }
      }
    }
  }
}